split integration tests into multiple files for batch and gui
